Manage product 5

+++ app/Http/Controllers/CategoryController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Category;
+use App\Models\Category;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
@@ -99,8 +99,8 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-        $category=Category::withCount(['child'])->find($id);
-        if($category->child_count==0){
+        $category=Category::withCount(['child','product'])->find($id);
+        if($category->child_count==0&&$category->prduct_count==0){
             $category->delete();
             return redirect()->route('category.index')->with(['success'=>'The category has been delete']);
         }
